co-locate normalize schema
