+++ libs/deepagents/src/middleware/utils.ts
@@ -7,47 +7,6 @@
 import { SystemMessage } from "@langchain/core/messages";
 
 /**
- * Recursively add `additionalProperties: false` to every object-typed node
- * in a JSON Schema.
- */
-export function normalizeSchema(
-  schema: Record<string, unknown>,
-): Record<string, unknown> {
-  if (schema.type !== "object" && schema.type !== "array") {
-    return schema;
-  }
-
-  if (schema.type === "array") {
-    const result: Record<string, unknown> = { ...schema };
-    const items = result.items;
-    if (items != null && typeof items === "object" && !Array.isArray(items)) {
-      result.items = normalizeSchema(items as Record<string, unknown>);
-    }
-    return result;
-  }
-
-  const result: Record<string, unknown> = {
-    ...schema,
-    additionalProperties: false,
-  };
-
-  const props = schema.properties;
-  if (props != null && typeof props === "object" && !Array.isArray(props)) {
-    const normalizedProps: Record<string, unknown> = {};
-    for (const [k, v] of Object.entries(props as Record<string, unknown>)) {
-      if (v != null && typeof v === "object" && !Array.isArray(v)) {
-        normalizedProps[k] = normalizeSchema(v as Record<string, unknown>);
-      } else {
-        normalizedProps[k] = v;
-      }
-    }
-    result.properties = normalizedProps;
-  }
-
-  return result;
-}
-
-/**
  * Append text to a system message.
  *
  * Creates a new SystemMessage with the text appended to the existing content.
